cleanup - avoid unnecessary boxing

+++ h2/src/main/org/h2/value/ValueShort.java
@@ -130,7 +130,7 @@ public class ValueShort extends Value {
 
     @Override
     public Object getObject() {
-        return Short.valueOf(value);
+        return value;
     }
 
     @Override
